fix: 채팅consumer 수정
