+++ src/main/java/nbc_final/gathering/domain/chatting/chatuser/entity/ChatMember.java
@@ -1,0 +1,24 @@
+package nbc_final.gathering.domain.chatting.chatuser.controller;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity(name = "MEMBER")
+public class ChatMember {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "member_id")
+    private Long id;
+
+    private String username;
+
+    public ChatMember(String username) {
+        this.username = username;
+    }
+}
