+++ src/main/java/nbc_final/gathering/domain/chatting/chatuser/repository/ChatMemberRepository.java
@@ -1,0 +1,2 @@
+package nbc_final.gathering.domain.chatting.chatuser.repository;public interface ChatMemberRepository {
+}
