--- src/main/java/nbc_final/gathering/domain/chatting/user/entity/ChatMember.java
@@ -1,40 +1,0 @@
-package nbc_final.gathering.domain.chatting.user.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.data.annotation.Id;
-import org.springframework.stereotype.Service;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity(name = "CHATMEMBER")
-public class ChatMember {
-
-    @jakarta.persistence.Id
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "chat_member_id")
-    private Long id;
-
-    private String username;
-
-    public ChatMember(String username) {
-        this.username = username;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Long getId() {
-        return id;
-    }
-}
